new commit

+++ src/main/java/com/example/demo/models/OperationRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface OperationRepository extends JpaRepository<Operation, Integer> {
     List<Operation> findAll();
-    Operation findOperationById(long itemId);
+    Operation findOperationById(int itemId);
 }
